Dividing by 0 with infinite ref sphere in waveabr.py
Correct an error in which parallel rays to the chief ray produced a dividing by 0 error with an infinite reference sphere in waveabr.py. This error can occur when using a ray grid with an odd side size, and so the central ray is the chief ray, which is parallel to itself.